Retire naming special-cases knowledge seam
